Cleaned up `/scala/server/` folder
