+++ app/src/main/scala/server/JsonEncoder.scala
@@ -1,7 +1,5 @@
 package jsonhandler
 
-import io.circe._
-
 import org.json4s._
 import org.json4s.JsonDSL._
 import org.json4s.jackson.JsonMethods._
@@ -11,28 +9,6 @@ import environment.cell._
 import environment.layer._
 import environment.element._
 import environment.generator.ElementSeeds._
-
-
-object Decoder {
-
-  def extractString(field: String, data: Json): Option[String] = {
-    val cursor: HCursor = data.hcursor
-    cursor.downField(field).as[String] match {
-      case Left(_)  => None
-      case Right(s) => Some(s)
-    }
-  }
-
-  def extractInt(field: String, data: Json): Option[Int] = {
-    val cursor: HCursor = data.hcursor
-    cursor.downField(field).as[Int] match {
-      case Left(_)  => None
-      case Right(i) => Some(i)
-    }
-  }
-
-}
-
 
 
 object Encoder {
@@ -48,6 +24,10 @@ object Encoder {
       (name -> m.map { case (k,v) => k -> v })
     return compact(render(json))
   }
+
+
+
+  // Type Specific encoders
 
   def encodeCell(c: Cell): String = {
     val json =
